Got basic memory output working
It doesn't update in a separate thread like the last commit does, unfortunately. However, it works.

More unfortunate, though, is that the memory values seem to either be static even when re-fetching them, or none of the memory stats being output reflect the memory usage the Tesla overlay may be using.

In either case, it's no help. I'll just abandon this branch, guess that it's running out of memory, and try to re-organize the GUI to address it because I don't know what else can be the problem.